Cleaned up signal to mqtt implementation; implemented edit amb
